requirements.txt
flask==2.0.1
opencv-python==4.5.3.56
torch==1.9.0
pandas==1.3.3
Pillow==8.3.2
facenet-pytorch==2.5.2
gspread==5.0.1
oauth2client==4.1.3
openpyxl==3.0.9 